add ioctl account balanceCmd
